+++ migrations/sqlite/2025-04-09-162257_custom_field/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS custom_field (
-    id INTEGER PRIMARY KEY NOT NULL,
-    provider_id INTEGER NOT NULL,
+    id BIGINT PRIMARY KEY NOT NULL,
+    provider_id BIGINT NOT NULL,
     field_name TEXT NOT NULL,
     field_type TEXT NOT NULL, -- e.g., 'text', 'integer', 'float', 'boolean', 'json'
     text_value TEXT,
@@ -8,6 +8,6 @@ CREATE TABLE IF NOT EXISTS custom_field (
     float_value REAL,
     boolean_value BOOLEAN,
     description TEXT,
-    created_at INTEGER NOT NULL,
-    updated_at INTEGER NOT NULL
+    created_at BIGINT NOT NULL,
+    updated_at BIGINT NOT NULL
 );
